sched/core: have io_schedule_prepare() return a long
In preparation for needing more state then 32-bit on 64-bit archs,
switch it to a long instead.

+++ block/blk-cgroup.c
@@ -1894,7 +1894,7 @@ static void blkcg_maybe_throttle_blkg(struct blkcg_gq *blkg, bool use_memdelay)
 	u64 now = blk_time_get_ns();
 	u64 exp;
 	u64 delay_nsec = 0;
-	int tok;
+	long tok;
 
 	while (blkg->parent) {
 		int use_delay = atomic_read(&blkg->use_delay);
